requirements.txt
fastapi
fastapi-sqlalchemy
pydantic
alembic
psycopg2
gunicorn
uvicorn
python-dotenv
SQLModel
pandas
opencv-python
openpyxl